fixed mixed precision usage. ~10s savings avoiding mixed calcs

+++ optimizers/muon.py
@@ -43,7 +43,7 @@ def zeropower_via_newtonschulz5(G: torch.Tensor, steps: int = 5) -> torch.Tensor
     """Newton-Schulz iteration to compute the zeroth power / orthogonalization of G."""
     assert G.ndim >= 2
     a, b, c = (3.4445, -4.7750, 2.0315)
-    X = G.half()
+    X = G.bfloat16()
 
     if G.size(-2) > G.size(-1):
         X = X.mT
